+++ node/clients/index.ts
@@ -1,8 +1,10 @@
 import { AuthType, ClientsConfig, IOClients, LRUCache } from '@vtex/api'
 import SchemaDataSource from './schemas'
+import DocumentsDataSource from './documents'
 import { forEachObjIndexed } from 'ramda'
 
 const memoryCache = {
+  documents: new LRUCache<string, any>({ max: 4000 }),
   schema: new LRUCache<string, any>({ max: 4000 }),
 }
 
@@ -13,6 +15,9 @@ forEachObjIndexed((cacheInstance: LRUCache<string, any>, cacheName: string) => {
 export class Clients extends IOClients {
   public get schemas(): SchemaDataSource {
     return this.getOrSet('schemas', SchemaDataSource)
+  }
+  public get documents(): DocumentsDataSource {
+    return this.getOrSet('documents', DocumentsDataSource)
   }
 }
 
@@ -26,6 +31,10 @@ export const clients: ClientsConfig<Clients> = {
     schemas: {
       authType: AuthType.bearer,
       memoryCache: memoryCache.schema,
+    },
+    documents: {
+      authType: AuthType.bearer,
+      memoryCache: memoryCache.documents,
     },
   },
 }