+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Add `saveDocuments` mutation.
+- Add `saveDocument` and `updateDocument` mutations.
 
 ## [0.3.0] - 2019-04-25